+++ Develappers.BillomatNet/InvoiceService.cs
@@ -436,5 +436,43 @@ namespace Develappers.BillomatNet
             var jsonModel = await GetItemByIdAsync<InvoiceCommentWrapper>($"/api/invoice-comments/{id}", token).ConfigureAwait(false);
             return jsonModel.ToDomain();
         }
+
+        /// <summary>
+        /// Creates an invoice comment.
+        /// </summary>
+        /// <param name="model">The invoice comment.</param>
+        /// <param name="token">The cancellation token.</param>
+        /// <returns>
+        /// A task that represents the asynchronous operation.
+        /// The task result returns the newly created invoice comment with the ID.
+        /// </returns>
+        /// <exception cref="ArgumentException">Thrown when the parameter check fails.</exception>
+        /// <exception cref="NotAuthorizedException">Thrown when not authorized to access this resource.</exception>
+        /// <exception cref="NotFoundException">Thrown when the resource url could not be found.</exception>
+        public async Task<InvoiceComment> CreateCommentAsync(InvoiceComment model, CancellationToken token = default)
+        {
+            if (model == null)
+            {
+                throw new ArgumentNullException(nameof(model));
+            }
+            if (model.InvoiceId == 0 || string.IsNullOrEmpty(model.Comment) || model.Id != 0)
+            {
+                throw new ArgumentException("invalid property values for invoice comment", nameof(model));
+            }
+            var wrappedModel = new InvoiceCommentWrapper
+            {
+                InvoiceComment = model.ToApi()
+            };
+            try
+            {
+                var result = await PostAsync("/api/invoice-comments", wrappedModel, token);
+                return result.ToDomain();
+            }
+            catch (WebException wex)
+                when (wex.Status == WebExceptionStatus.ProtocolError && (wex.Response as HttpWebResponse)?.StatusCode == HttpStatusCode.BadRequest)
+            {
+                throw new ArgumentException("wrong input parameter", nameof(model), wex);
+            }
+        }
     }
 }