+++ Develappers.BillomatNet/Mapping/MappingExtensions.cs
@@ -152,6 +152,11 @@ namespace Develappers.BillomatNet.Mapping
             return s_invoiceCommentMapper.ApiToDomain(value);
         }
 
+        internal static Api.InvoiceComment ToApi(this InvoiceComment value)
+        {
+            return s_invoiceCommentMapper.DomainToApi(value);
+        }
+
         internal static Types.PagedList<Contact> ToDomain(this ContactListWrapper value)
         {
             return s_contactMapper.ApiToDomain(value);